This fixes a bug where an exception during delete might return a 500
This fixes a bug where if an exception was thrown for the top level path being deleted, all recorded errors would be ignored and a 500 would be thrown because the exception would not be caught correctly. This fixes that so that when delete is called on the top level path and an exception is raised we record that exception (like we do for the other exceptions) and correctly return the exception with the correct HTTP status code rather than always a 500.